Added project source dir to octave path.
Added a variable to launch.json to set the source directory so to add to octave's path every time you debug.
That way the source can be located anywhere on disk. Also added an option to set octave's executable, in case you don't have it on your path. Updated documentation.