Refactored the matrix fetch code to be more efficient.
Instead of always getting all the elements, now I get at most 100 elements, unless specified otherwise in the launch.json via prefetchCount.
It still takes a second to fetch the 100 elements requested by vscode. I plan to change that to fetch them in one single call and parse the values.
I've tested a partial implementation, and it's definitely faster. But the parsing makes the code more error prone and easy to break.