Fixed a bug in the parsing of multi-column vectors.
Octave will output different things depending on the number of columns a vector has.
For vectors with more elements than the amount that can fit one line, the remaining elements are output in the next line.
Depending if there are 1, 2 or more, octave will output e.g.
Column 11:
Columns 11 and 12:
Columns 1 through 5:
Added a regex that supports those three cases.