Implemented parsing for SparseMatrix.
On issue with the current approach is that fetching children fetches indices needed to name the children.
However, if the value is valid from the beginning, the indices need to be fetched upon construction.
Currently this is done on loadNew but since this is not done in the constructor the user might forget to do it.