Create an UnknownType, and make sure of its size allows loading.
For very large scalars, e.g. unknown matrix types, the variable request might take so much time that by the time it finishes the variable might not exist anymore if the person has sent a step by that time. Because the commands are interpreted faster than the dbg input can be processed. The UnknownType fixes this issue.