Added get raw value, to skip removing ans et al.
When processing strings we can't just remove all white spaces, as these might be part of the string value. So I added a getRawValue for that case.
In practice it would make more sense to use that function all the time, and in each location where the function is called remove only the required prefix.